Add common methods for implementations to the `E2EUser`.
